rename relatedContent => onwardsContent
